+++ src/strftime.c
@@ -186,7 +186,7 @@ static int add_num_time_t(char *string, int max, time_t num)
 	if (sizeof(num) > 16)
 		abort();
 	length = snprintf(buf, sizeof(buf), "%lu", (unsigned long)num);
-	assert(length >= 0 && length<sizeof(buf));
+	assert(length >= 0 && (size_t)length<sizeof(buf));
 	length = add_str(string, buf, max);
 	return length;
 }
